requirements.txt
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.12.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
whitenoise==6.1.0
gunicorn==20.0.4